requirements.txt
django>=3.0,<4.0
django_rest_framework==0.1.0
celery<6.0,>=5.0
redis==3.5.3
drf-yasg==1.20.0
django-rest-swagger==2.2.0
django_extensions==2.1.9
django-choices==1.7.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
coverage==5.3
pytest==6.1.1
pytest-cov==2.10.1
factory-boy==2.12.0
pytest-factoryboy==2.0.3
pytest-django==4.0.0
responses==0.15.0
